--- projet/urls.py
@@ -1,8 +1,0 @@
-from django.urls import path, include
-from . import views
-
-urlpatterns = [
-    path('projet/', views.show_home),
-    path('new_project/', views.new_project, name="new_project"),
-    path('login/', views.check_login, name="login_projet")
-]
